+++ resources/views/admin/UserManagement/Role/index.blade.php
@@ -5,7 +5,9 @@
         <div class="card">
             <div class="card-header">
                 <h3 class="card-title">Role Table</h3>
-                <a class="card-tools" href="{{route('role.create')}}"><i class="fas fa-plus"></i></a>
+                @can('create',new app\Models\Role())
+                    <a class="card-tools" href="{{route('role.create')}}"><i class="fas fa-plus"></i></a>
+                @endcan
             </div>
             <div class="card-body">
                 <table class="table table-bordered">
@@ -13,7 +15,9 @@
                         <tr>
                             <th style="width: 10px">#</th>
                             <th>Name</th>
-                            <th>Action</th>
+                            @canany(['update','delete'],new app\Models\Role())
+                                <th>Action</th>
+                            @endcanany
                         </tr>
                     </thead>
                     <tbody>
@@ -21,10 +25,16 @@
                         <tr>
                             <td>{{$loop->iteration}}</td>
                             <td>{{ucFirst($role->name)}}</td>
-                            <td>
-                                <a href="{{route("role.edit",['role'=>$role->id])}}"><i class="fas fa-edit"></i></a>&nbsp;
-                                <a onclick='deleteItem("{{route("role.destroy",["role"=>$role->id])}}")'><i class="fas fa-trash text-red"></i></a>
-                            </td>
+                            @canany(['update','delete'],new app\Models\Role())
+                                <td>
+                                    @can('update',new app\Models\Role())
+                                        <a href="{{route("role.edit",['role'=>$role->id])}}"><i class="fas fa-edit"></i></a>&nbsp;
+                                    @endcan
+                                    @can('delete',new app\Models\Role())
+                                        <a onclick='deleteItem("{{route("role.destroy",["role"=>$role->id])}}")'><i class="fas fa-trash text-red"></i></a>
+                                    @endcanany
+                                </td>
+                            @endcanany
                         </tr>
                         @endforeach
                     </tbody>
